+++ src/main/scala/jukaiScala/merlin/Transpose.scala
@@ -8,8 +8,8 @@ object Transpose extends Functor{
 
   override val functorName: String = "Transpose"
 
-  override def convert(data: DenseMatrix[Double]): DenseMatrix[Double] = data.t
+  override def convert(data: DenseMatrix[Float]): DenseMatrix[Float] = data.t
 
-  def apply(m: DenseMatrix[Double]) = this.convert(m)
+  def apply(m: DenseMatrix[Float]) = this.convert(m)
 
 }
